requirements.txt
pygame==2.6.1
requests==2.32.3
pillow==11.1.0
pycparser==2.22
cffi==1.17.1
wifi==0.3.8
pyinstaller
speedtest-cli
